Backend/envios/requirements.txt
fastapi
uvicorn
pydantic
kafka-python
requests